更新数据库字段

+++ src/main/java/com/atxbai/online/model/pojo/Manager.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class Manager {
     @TableId(value = "no",type = IdType.AUTO)
-    private Long no;
+    private Integer no;
 
     @TableField("name")
     private String name;
